+++ lib/extensions/extensions.dart
@@ -1,0 +1,9 @@
+library all_extensions;
+
+import 'package:easy_localization/easy_localization.dart';
+import 'package:flutter/material.dart';
+
+part 'context_extension.dart';
+part 'string_extension.dart';
+part 'text_style_extension.dart';
+part 'widget_extension.dart';