+++ lib/extensions/widget_extension.dart
@@ -1,0 +1,5 @@
+part of 'extensions.dart';
+
+extension WidgetExtension on Widget {
+  SliverToBoxAdapter get toSliverToBoxAdapter => SliverToBoxAdapter(child: this);
+}